Add alternating optimization and count strategy tooling
- add alternating RL/OLS optimization with resumable checkpoints, constrained regression modes, graphing, and edge evaluation\n- add count-strategy comparison tooling and unified per-run logging\n- fix Kelly growth evaluation behavior and update project documentation

+++ src/Game/Hand.cpp
@@ -116,7 +116,7 @@ Hand Hand::split() {
     }
     
     // Save the original bet and cards
-    unsigned int originalBet = bet;
+    double originalBet = bet;
     Card firstCard = cards[0];
     Card secondCard = cards[1];
